fix(desktop): ship a full hicolor icon set so Linux resolves the app icon
## Problem

The Linux packages installed exactly one app icon, at
`hicolor/1024x1024/apps/traycer-desktop.png`. The freedesktop icon theme
spec only searches size directories declared in the theme's
`index.theme`, and stock hicolor stops at 512x512 — so no icon lookup
ever visited that directory. The generated `.desktop` entry's
`Icon=traycer-desktop` key silently resolved to nothing and the shell
fell back to a generic icon in the dock, app grid and window switcher.

`build.linux` set no `icon`, so electron-builder fell back to the
1024x1024 `buildResources/icon.png` and emitted a single entry at the
source's own dimensions.

## Summary

- point `build.linux.icon` at a generated `resources/bundle/icons/` set
  covering sizes hicolor actually declares (16, 24, 32, 48, 64, 128,
  256, 512)
- add `scripts/assets/generate-linux-icons.cjs`, which downscales that
  set from `icon.png` with a box filter (a 2-tap bilinear sample reads 4
  of the 4096 source pixels behind each 1024 -> 16 output pixel and
  aliases badly). Dependency-free, mirroring `generate-tray-icons.cjs`
- extract the PNG codec the two generators would otherwise duplicate
  into `scripts/assets/png-codec.cjs`
- `check-bundle-icons.cjs` now fails packaging when a Linux size is
  missing or a filename disagrees with its IHDR dimensions, which would
  install the icon under the wrong hicolor size

## Verification

- laid the generated set out exactly as electron-builder installs it and
  resolved it with GTK against the real system `index.theme`: all of 16,
  22, 24, 32, 48, 64, 96, 128, 256, 512 resolve, against not-found at
  every size beforehand
- the codec extraction is behaviour-preserving: the tray generator's
  output is byte-identical to the pre-refactor script's. (Neither
  reproduces the committed tray PNGs byte-for-byte — they differ by 1-4
  bytes of compression only. That drift predates this change and comes
  from a different zlib build; the committed tray assets are untouched
  here.)
- `check-bundle-icons.cjs` passes, and was confirmed to fail on both a
  removed size and a filename/IHDR mismatch

Note this fixes future packages only; existing installs keep the generic
icon until reinstalled.
